+++ src/main/java/com/example/bookstorejdbc/service/CategoryService.java
@@ -1,12 +1,14 @@
 package com.example.bookstorejdbc.service;
 
 import com.example.bookstorejdbc.data.dto.CategoryDto;
+import com.example.bookstorejdbc.data.entity.Book;
 import com.example.bookstorejdbc.data.entity.Category;
 import com.example.bookstorejdbc.data.mapper.CategoryMapper;
 import com.example.bookstorejdbc.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,8 +25,8 @@ public class CategoryService {
         return repository.findAll().stream().map(mapper::toDto).collect(Collectors.toList());
     }
 
-    public CategoryDto getById(Integer id) {
-        return mapper.toDto(repository.findById(id));
+    public Optional<CategoryDto> getById(Integer id) {
+        return repository.findById(id).map(mapper::toDto);
     }
 
     public void addNewCategory(CategoryDto category) {
@@ -32,13 +34,18 @@ public class CategoryService {
     }
 
     public void deleteCategory(Integer id) {
-        repository.deleteById(id);
+        Optional<Category> сategoryOptional = repository.findById(id);
+        if (сategoryOptional.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Категория с id " + id + " не найдена");
+        }
     }
 
     public void updateCategory(CategoryDto category, Integer id) {
-        Category oldCategory = repository.findById(id);
-        Category newCategory = mapper.toEntity(category);
-        newCategory.setCategory_id(oldCategory.getCategory_id());
-        repository.update(newCategory);
+        Category entity = repository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Категория с id " + id + " не найдена"));
+        entity.setName(category.getName());
+        repository.update(entity);
     }
 }
