+++ src/main/java/com/example/bookstorejdbc/service/PublishingHouseService.java
@@ -1,12 +1,15 @@
 package com.example.bookstorejdbc.service;
 
 import com.example.bookstorejdbc.data.dto.PublishingHouseDto;
+import com.example.bookstorejdbc.data.entity.Book;
+import com.example.bookstorejdbc.data.entity.Category;
 import com.example.bookstorejdbc.data.entity.PublishingHouse;
 import com.example.bookstorejdbc.data.mapper.PublishingHouseMapper;
 import com.example.bookstorejdbc.repository.CrudRepository;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,8 +26,8 @@ public class PublishingHouseService {
         return repository.findAll().stream().map(mapper::toDto).collect(Collectors.toList());
     }
 
-    public PublishingHouseDto getById(Integer id) {
-        return mapper.toDto(repository.findById(id));
+    public Optional<PublishingHouseDto> getById(Integer id) {
+        return repository.findById(id).map(mapper::toDto);
     }
 
     public void addNewPublishingHouse(PublishingHouseDto publishingHouse) {
@@ -32,13 +35,18 @@ public class PublishingHouseService {
     }
 
     public void deletePublishingHouse(Integer id) {
-        repository.deleteById(id);
+        Optional<PublishingHouse> publishingHouseOptional = repository.findById(id);
+        if (publishingHouseOptional.isPresent()) {
+            repository.deleteById(id);
+        } else {
+            throw new IllegalArgumentException("Книга с id " + id + " не найдена");
+        }
     }
 
     public void updatePublishingHouse(PublishingHouseDto publishingHouse, Integer id) {
-        PublishingHouse oldPublishingHouse = repository.findById(id);
-        PublishingHouse newPublishingHouse = mapper.toEntity(publishingHouse);
-        newPublishingHouse.setPublishing_house_id(oldPublishingHouse.getPublishing_house_id());
-        repository.update(newPublishingHouse);
+        PublishingHouse entity = repository.findById(id)
+                .orElseThrow(() -> new IllegalArgumentException("Издание с id " + id + " не найдено"));
+        entity.setName(publishingHouse.getName());
+        repository.update(entity);
     }
 }
